requirements.txt
dj-database-url
dj-static
gunicorn
psycopg2
static
django_jasmine
lettuce
django_nose
South
django_compressor
BeautifulSoup
djangorestframework
markdown
pyyaml
